Static Typing: Remove "basic types", all Blockly.Types define a unique type.
Now all Blockly.Types can be used indiscriminately.
The original basicTypes were introduced to maintain some level of compatibility with the original Blockly blocks type assignment for block outputs and checks. Since all the blocks have to be modified anyway, it does not make sense to keep this old rule.