Update board manager to contain compiler flag info, and reorganise board profiles.
This commit reorganises a little bit the arduino board profiles.
It also adds an additioanl property to the profile to include the IDE CLI board flag.
This is currently not used, as it is stored and managed by the server, but this is the
first step to move that to the front end, as it makes more sense to keep that data
together with the rest of the board profile.