First step for "instances", acceptably working, but not finished.
Instances are to be used for Blocks that need to create an "instance" of something that can then be used for creating custom-made field dropdowns.

Updated with the stepper block, as it was using a similar concept, but now can be used by any blocks.

Still need to fix some bugs with the "unique naming" from flyout to the workspace. And cleanup the code, comment documentation, etc.