Transition server and handlers to use new Bottle implementation
Moving from SimpleHTTPServer to a Bottle + Waitress implementation of the
Ardublockly Server.
The request have been changed slightly to be more "RESTful" (but not
really) and te actions module had to be updated accordingly.
The front end JavaScript had to be updated as well to work with the new
version.